refactor: extract CLI layer into deckslots.cli subpackage
Move cli.py → cli/parser.py, commands.py → cli/commands.py,
repl.py → cli/repl.py, status.py → cli/status.py. Add cli/__init__.py
re-exporting ParsedCommand, parse_command, and main so existing
imports (`from deckslots.cli import ...`) and the
`deckslots.cli:main` entry point keep working unchanged.

Internal imports inside the new subpackage now point to the new
locations; loggers updated to `deckslots.cli.commands` /
`deckslots.cli.repl`.

Closes #72.
